backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
google-genai>=0.8.0
elevenlabs>=1.2.0
pydantic[email]>=2.0.0
motor>=3.4.0                     # async MongoDB driver
bcrypt>=4.1.0                    # password hashing
redis>=5.0.0                     # optional — only needed if using Redis sessions
